requirements.txt
tweepy>=4.14.0
feedparser>=6.0.10
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.1
